# rick-morty-tournament/api/openapi.yaml
openapi: 3.1.0
info:
  title: Rick and Morty Tournament API
  version: 0.1.0
  description: >
    Especificacao inicial da API proprietaria responsavel por catalogo,
    sincronizacao, enriquecimento e roster de torneio.
servers:
  - url: http://localhost:8080
paths:
  /v1/health:
    get:
      summary: Healthcheck da API
      responses:
        "200":
          description: API saudavel
  /v1/characters:
    get:
      summary: Lista personagens do catalogo local
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: eligible
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de personagens
  /v1/characters/{id}:
    get:
      summary: Detalha um personagem
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Personagem encontrado
        "404":
          description: Personagem nao encontrado
  /v1/characters/random:
    get:
      summary: Retorna personagens aleatorios
      parameters:
        - in: query
          name: count
          schema:
            type: integer
            minimum: 1
            maximum: 64
      responses:
        "200":
          description: Selecionados com sucesso
  /v1/tournament/rosters:
    post:
      summary: Gera roster do torneio
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                count:
                  type: integer
                  default: 16
                seed:
                  type: string
                strategy:
                  type: string
                  default: random_uniform
      responses:
        "201":
          description: Roster criado
  /v1/tournament/rosters/{id}:
    get:
      summary: Busca um roster existente
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Roster encontrado
        "404":
          description: Roster nao encontrado
  /v1/admin/sync/full:
    post:
      summary: Dispara sincronizacao completa
      responses:
        "202":
          description: Sync disparado
  /v1/admin/sync/runs:
    get:
      summary: Lista execucoes de sincronizacao
      responses:
        "200":
          description: Lista de sync runs
  /v1/admin/enrichment/run:
    post:
      summary: Dispara enriquecimento em lote
      responses:
        "202":
          description: Enriquecimento disparado
  /v1/admin/enrichment/runs:
    get:
      summary: Lista execucoes de enriquecimento
      responses:
        "200":
          description: Lista de enrichment runs